+++ sparkle/src/components/markdown/index.ts
@@ -3,3 +3,5 @@ export * from "./CodeBlockWithExtendedSupport";
 export * from "./ContentBlockWrapper";
 export * from "./Markdown";
 export * from "./MarkdownContentContext";
+export * from "./TableBlock";
+export * from "./utils";
