+++ sparkle/src/components/markdown/utils.ts
@@ -1,0 +1,41 @@
+export function sanitizeContent(str: string): string {
+  // (1) Add closing backticks if they are missing such that we render a code block or inline
+  // element during streaming.
+
+  // Regular expression to find either a single backtick or triple backticks
+  const regex = /(`{1,3})/g;
+  let singleBackticks = 0;
+  let tripleBackticks = 0;
+
+  // Search for all backticks in the string and update counts
+  let match;
+  while ((match = regex.exec(str)) !== null) {
+    if (match[1] === "```") {
+      tripleBackticks++;
+    } else if (match[1] === "`") {
+      singleBackticks++;
+    }
+  }
+  // Append closing backticks if needed
+  if (tripleBackticks % 2 !== 0) {
+    if (str.endsWith("`")) {
+      str += "``";
+    } else if (str.endsWith("``")) {
+      str += "`";
+    } else {
+      str += str.includes("\n") ? "\n```" : "```";
+    }
+  } else if (singleBackticks % 2 !== 0) {
+    str += "`";
+  }
+
+  return str;
+}
+
+export function detectLanguage(children: React.ReactNode) {
+  if (Array.isArray(children) && children[0]) {
+    return children[0].props.className?.replace("language-", "") || "text";
+  }
+
+  return "text";
+}
